Added AutoMappper And Delete image on Post Edit/Delete
Delete  Functionality On Edit And Delete  Controller image delete from Images Folder.
Add Preview Image On edit Functionality.
Implement Mpping Between ViewModel And entity using AutoMapper

+++ src/Presentation/BlogApp.Web/Controllers/PostController.cs
@@ -1,11 +1,14 @@
 ﻿using System.Security.Claims;
+using AutoMapper;
 using BlogApp.Core.Repositories;
 using BlogApp.Core.Services;
 using BlogApp.Models;
 using BlogApp.ViewModel;
+using BlogApp.ViewModels;
 using BlogApp.Web.Helper;
 using BlogApp.Web.ViewModel;
 using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Http.HttpResults;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.Rendering;
 
@@ -14,23 +17,35 @@ namespace BlogApp.Controllers
     [Route("[Controller]")]
     public class PostController : Controller
     {
+        #region Private Fields
+
         private readonly IPostService _postService;
         private readonly ICommentService _commentService;
         private readonly ILikeService _likeService;
         private readonly IGenericRepository<Category> _categoryRepository;
+        private readonly IMapper _mapper;
+
+        #endregion
+
+        #region Contructor
 
         public PostController(
             IPostService postService,
             ICommentService commentService,
             ILikeService likeService,
+            IMapper mapper,
             IGenericRepository<Category> categoryRepository)
         {
             _postService = postService;
             _commentService = commentService;
             _likeService = likeService;
+            _mapper = mapper;
             _categoryRepository = categoryRepository;
         }
 
+        #endregion
+
+        #region Public Methods
 
         [HttpGet("")]
         public async Task<IActionResult> Index(
@@ -43,7 +58,7 @@ namespace BlogApp.Controllers
             string userId = User.FindFirstValue(ClaimTypes.NameIdentifier) ?? string.Empty;
             bool isAdmin = User.IsInRole("Admin");
 
-            var result = await _postService.GetPostsAsync(
+            var (posts, totalCount) = await _postService.GetPostsAsync(
                 categoryId,
                 is_most_Liked,
                 sortOrder,
@@ -52,27 +67,50 @@ namespace BlogApp.Controllers
                 userId,
                 isAdmin);
 
-            ViewBag.Categories = _categoryRepository.GetAll().ToList();
-            ViewBag.CurrentPage = page;
-            ViewBag.TotalPages = (int)Math.Ceiling((double)result.totalCount / pageSize);
-            ViewBag.CategoryId = categoryId;
-            ViewBag.SortOrder = sortOrder;
-            ViewBag.IsMostLiked = is_most_Liked;
+            var postsListViewModel = new PostsListViewModel
+            {
+                CurrentPage = page,
+                TotalPages = (int)Math.Ceiling((double)totalCount / pageSize),
+                SortOrder = sortOrder,
+                IsMostLiked = is_most_Liked,
+                CategoryId = categoryId,
+                Categories = _mapper.Map<List<CategoryViewModel>>(_categoryRepository.GetAll().ToList()),
+                Posts = _mapper.Map<List<PostViewModel>>(posts)
+            };
 
-            return View(result.posts);
+            return View(postsListViewModel);
         }
-
 
         [HttpGet("Detail")]
         public async Task<IActionResult> Detail(int id)
         {
-            var post = await _postService.GetPostDetailAsync(id);
-
-            if (post == null)
+            if (id <= 0)
                 return NotFound();
 
-            return View(post);
+            string userId = User.FindFirstValue(ClaimTypes.NameIdentifier);
+
+            bool isAdmin = User.IsInRole("Admin");
+
+            var post = await _postService.GetPostDetailAsync(id);
+
+            var comments = await _commentService.GetCommnetsByPostIdAsync(id);
+
+            var commentViewModel = comments.Select(c => new CommentViewModel
+            {
+                Id = c.Id,
+                Content = c.Content,
+                CommentDate = c.CommentDate,
+                UserName = c.User?.UserName ?? "",
+                PostId = c.PostId,
+                CanDelete = isAdmin || c.UserId == userId || c.Post.CreatedByUserId == userId
+            }).ToList();
+
+            var postVM = _mapper.Map<PostViewModel>(post);
+            postVM.Comments = commentViewModel;
+
+            return View(postVM);
         }
+
 
         [HttpGet("Create")]
         [Authorize(Roles = "Admin,Author")]
@@ -95,7 +133,13 @@ namespace BlogApp.Controllers
 
             try
             {
-                var post = vm.ToDataModel(vm);
+                #region Upload Image
+
+                ImagesHelper.UploadFile(vm.FeatureImage);
+
+                #endregion
+
+                var post = _mapper.Map<Post>(vm);
 
                 await _postService.CreatePostAsync(post);
 
@@ -109,29 +153,7 @@ namespace BlogApp.Controllers
             }
         }
 
-        //[Authorize(Roles = "Admin,Author")]
-        //[HttpGet("Edit")]
-        //public async Task<IActionResult> Edit(int id)
-        //{
-        //    var post = await _postService.GetPostDetailAsync(id);
 
-        //    if (post == null)
-        //        return NotFound();
-
-        //    var vm = new PostViewModel()
-        //    {
-        //        Id = id,
-        //        Title    = post.Title,
-        //        Content  = post.Content,
-        //        CategoryId  = post.CategoryId,
-        //        Categories = LoadCategories(),
-        //        Authore = post.Authore,
-        //        FeatureImage = ImagesHelper.GetImage(post.FeatureImagePath),
-        //        PublishedDate = post.PublishedDate
-        //    };
-
-        //    return View(vm);
-        //
         [Authorize(Roles = "Admin,Author")]
         [HttpGet("Edit")]
         public async Task<IActionResult> Edit(int id)
@@ -141,19 +163,8 @@ namespace BlogApp.Controllers
             if (post == null)
                 return NotFound();
 
-            var vm = new PostViewModel
-            {
-                Id = post.Id,
-                Title = post.Title,
-                Content = post.Content,
-                Authore = post.Authore,
-                CategoryId = post.CategoryId,
-                FeatureImagePath = post.FeatureImagePath,
-                CreatedByUserId = post.CreatedByUserId,
-                Categories = LoadCategories()
-            };
-
-            //LoadCategories();
+            var vm = _mapper.Map<PostViewModel>(post);
+            vm.Categories = LoadCategories();
 
             return View(vm);
         }
@@ -173,15 +184,18 @@ namespace BlogApp.Controllers
             if (post == null)
                 return NotFound();
 
-            post.Title = vm.Title;
-            post.Content = vm.Content;
-            post.Authore = vm.Authore;
-            post.CategoryId = vm.CategoryId;
 
-            if (vm.FeatureImage != null)
+            if (vm.FeatureImage != null && vm.FeatureImage.Length > 0)
             {
+
+                ImagesHelper.DeleteFile(post.FeatureImagePath);
+
                 post.FeatureImagePath = ImagesHelper.UploadFile(vm.FeatureImage);
             }
+
+            post.Title = vm.Title;
+            post.Content = vm.Content;
+            post.CategoryId = vm.CategoryId;
 
             await _postService.EditPostAsync(post);
 
@@ -193,6 +207,7 @@ namespace BlogApp.Controllers
         public async Task<IActionResult> Delete(int id)
         {
             var post = await _postService.GetPostDetailAsync(id);
+
             if (post == null)
                 return NotFound();
 
@@ -204,7 +219,14 @@ namespace BlogApp.Controllers
         {
             try
             {
+                var post = await _postService.GetPostDetailAsync(id);
+                if (post == null)
+                    return NotFound();
+
                 await _postService.DeletePostAsync(id);
+
+                ImagesHelper.DeleteFile(post.FeatureImagePath);
+
                 TempData["SuccessMessage"] = "Post deleted successfully!";
             }
             catch (Exception ex)
@@ -215,39 +237,62 @@ namespace BlogApp.Controllers
             return RedirectToAction("Index");
         }
 
-
         [HttpPost("AddComment")]
         [Authorize]
-        public async Task<IActionResult> AddComment([FromBody] Comment comment)
+        public async Task<IActionResult> AddComment([FromBody] CommentViewModel model)
         {
+            string userId = User.FindFirstValue(ClaimTypes.NameIdentifier);
+            bool isAdmin = User.IsInRole("Admin");
 
-            string? userName = User.Identity?.Name;
+
+            var comment = _mapper.Map<Comment>(model);
+            comment.UserId = userId;
+
+            await _commentService.AddCommentAsync(comment);
+
+            var comments = await _commentService.GetCommnetsByPostIdAsync(model.PostId);
 
 
-            if (string.IsNullOrWhiteSpace(userName))
+            var commentVM = _mapper.Map<List<CommentViewModel>>(comments);
+
+            foreach (var c in commentVM)
             {
+                var original = comments.First(x => x.Id == c.Id);
 
-                return Forbid();
+                c.CanDelete = isAdmin
+                    || original.UserId == userId
+                    || original.Post.CreatedByUserId == userId;
             }
 
-
-            await _commentService.AddCommentAsync(comment, userName!);
-
-
-            return Json(new
-            {
-                username = comment.UserName,
-                content = comment.Content,
-                commentDate = comment.CommentDate.ToString("dd MMM yyyy")
-            });
+            return PartialView("PostCommentsList", commentVM);
         }
-
         [HttpPost("DeleteComment")]
-        [Authorize(Roles = "Admin,Author")]
-        public async Task<IActionResult> DeleteComment(int id)
+        [Authorize]
+        public async Task<IActionResult> DeleteComment(int id, int postId)
         {
-            await _commentService.DeleteCommentAsync(id);
-            return Ok();
+            string userId = User.FindFirstValue(ClaimTypes.NameIdentifier);
+            bool isAdmin = User.IsInRole("Admin");
+
+            var deleted = await _commentService.DeleteCommentAsync(id, userId, isAdmin);
+
+            if (!deleted)
+                return Forbid();
+
+            var comments = await _commentService.GetCommnetsByPostIdAsync(postId);
+
+
+            var commentVM = _mapper.Map<List<CommentViewModel>>(comments);
+
+            foreach (var c in commentVM)
+            {
+                var original = comments.First(x => x.Id == c.Id);
+
+                c.CanDelete = isAdmin
+                    || original.UserId == userId
+                    || original.Post.CreatedByUserId == userId;
+            }
+
+            return PartialView("PostCommentsList", commentVM);
         }
 
         [HttpPost("ToggleLike")]
@@ -272,7 +317,6 @@ namespace BlogApp.Controllers
                 count = result.count
             });
         }
-
 
         [Authorize(Roles = "Admin,Author")]
         [HttpPost("TogglePublish")]
@@ -300,18 +344,9 @@ namespace BlogApp.Controllers
             }
         }
 
+        #endregion
 
-        //private IEnumerable<SelectListItem> LoadCategories()
-        //{
-        //    var categories = _categoryRepository.GetAll().ToList();
-
-        //    var selectListItemsList = categories.Select(c => new SelectListItem
-        //    {
-        //        Value = c.Id.ToString(),
-        //        Text = c.Name
-        //    }).ToList();
-        //    return selectListItemsList;
-        //}
+        #region Private Methods
 
         private IEnumerable<SelectListItem> LoadCategories()
         {
@@ -323,5 +358,6 @@ namespace BlogApp.Controllers
                 });
         }
 
+        #endregion
     }
 }
